Adding side option to offsetSingleGroup
The new option side allow controlling the direction the data points are offset to. The default is side=0 for shifting in both direction. 1 shifts data points only to positive values, i.e. right or top, while -1 shifts to negative values. For other values only the sign is considered without throwing a warning.